+++ vpc-20160428/ChangeLog.md
@@ -1,3 +1,9 @@
+2025-07-24 Version: 6.11.9
+- Update API AssociateVpcCidrBlock: add request parameters Ipv6CidrMask.
+- Update API CreateVpc: add request parameters Ipv6CidrMask.
+- Update API CreateVpc: add request parameters Ipv6IpamPoolId.
+
+
 2025-07-15 Version: 6.11.8
 - Update API DescribeNatGatewayAssociateNetworkInterfaces: add response parameters Body.AssociateNetworkInterfaces.$.ResourceVpcId.
 - Update API GetVpcPrefixListAssociations: add response parameters Body.PrefixListAssociation.$.CidrList.